Added backwards compatibility for locating provisioning stuff in legacy Xamarin folders
Dev17 still generates provisioning information in the legacy Xamarin paths, so we need a way for this code to consider both new and legacy paths.
This fixes an issue where Dev17 fails to build with newer iOS SDK versions that relies only on new MAUI paths.
The reason of the failure is because Dev17 uses legacy paths and the iOS SDK looks in new paths.

Fixes P0 bug #2505126